pyproject.toml
[project]
name = "kedro-graphql"
authors = [
    {name = "opensean"}
]
description = "A kedro plugin for serving any kedro project as a GraphQL api"
requires-python = ">=3.10, <3.12"
keywords = [
    "pipelines",
    "machine learning",
    "data pipelines",
    "data science",
    "data engineering",
    "graphql"
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dynamic = ["version", "readme"]
dependencies = [
    "kedro==0.19.11",
    "kedro-datasets>=3.0.0",
    "fastapi>=0.115.11",
    "pymongo>=4.7.2",
    "python-dotenv>=1.0.1",
    "gql[all]>=3.5.1",
    "strawberry-graphql>=0.262.5",
    "uvicorn[standard]>=0.30.0",
    "redis>=5.0.0",
    "click>=4.0",
    "celery>=5.4.0",
    "boto3>=1.26.160",
    "asyncio>=3.4.3",
    "PyJWT>=2.10.1",
    "python-multipart>=0.0.20",
]

[project.scripts]
kedro-graphql = "kedro_graphql.__main__:main"

[project.entry-points."kedro.project_commands"]
kedro-graphql = "kedro_graphql.commands:commands"

[project.entry-points."kedro.hooks"]
kedro-graphql-validation = "kedro_graphql.hooks:validation_hooks"
kedro-graphql-logging = "kedro_graphql.hooks:logging_hooks"

[project.entry-points."panel.auth"]
pkce = "kedro_graphql.ui.auth:PKCELoginHandler"

[project.urls]
Homepage = "https://github.com/opensean/kedro-graphql"
Source = "https://github.com/opensean/kedro-graphql"

[project.optional-dependencies]
test = [
    "pandas>=2.2.0",
    "pytest-cov>=3,<7",
    "pytest-mock>=1.7.1, <4.0",
    "pytest>=7.2,<9.0",
    "pytest-asyncio>=0.23.7",    
    "celery[pytest]~=5.4.0",
]
experimental = [
    "minio~=7.1.15", 
]
ui = [
    "panel>=1.6.1", 
    "kedro-viz>=10.2.0", 
    "sh>=2.2.2",
]

[tool.setuptools.dynamic]
readme = {file = "README.md", content-type = "text/markdown"}
version = {attr = "kedro_graphql.__version__"}

[tool.kedro]
package_name = "kedro_graphql"
project_name = "kedro-graphql"
kedro_init_version = "0.19.6"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = """
--cov-report term-missing \
--cov src/kedro_graphql -ra"""

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.kedro_telemetry]
project_id = "a359b3866d7b42889afff83354405d28"
